PICARD-1523: Compare preserved tags case-insensitive
